Emoji support update.

+++ include/lgi/common/Css.h
@@ -312,7 +312,7 @@ public:
 		AttachmentFixed,
 	};
 	
-	static double FontSizeTable[7]; // SizeXSmall etc
+	static float FontSizeTable[7]; // SizeXSmall etc
 
 	#define _FloatAbs(a) \
 		(((a) < 0.0f) ? -(a) : (a))
